bugfixex, errormsg and search

+++ src/components/DeleteAuctionBtn.jsx
@@ -1,19 +1,40 @@
 import React, { useState, useEffect } from "react";
-import { Button, Modal, Form } from "react-bootstrap";
+import { Button, Modal, Form, Alert } from "react-bootstrap";
 import { useAuctionApi } from "../AuctionApi";
 
-export default function DeleteAuctionBtn({ auction, onDelete }) {
+export default function DeleteAuctionBtn({ auction, onDelete, leadingBids }) {
   const { deleteAuction } = useAuctionApi();
   const [displayModal, setDisplayModal] = useState(false);
   const [deleteKey, setDeleteKey] = useState("");
+  const [error, setError] = useState("");
+  const [disableInput, setDisableInput] = useState(false);
 
+  // useEffect to check if any bids, disable input-fields.
+  useEffect(() => {
+    if (leadingBids && leadingBids.Amount) {
+      setDisableInput(true);
+    } else {
+      setDisableInput(false);
+    }
+  }, [leadingBids]);
+
+  // function to handle delete of auction
   const handleDelete = () => {
+    if (!deleteKey) {
+      // display error if "CreatedBy-value" is missing or wrong.
+      setError("Ange ditt hemliga token för att radera auktionen.");
+
+      setTimeout(() => {
+        setError("");
+      }, 3000);
+      // cancel function if token is missing
+      return;
+    }
+
     if (deleteKey === auction.CreatedBy) {
       deleteAuction(auction.AuctionID);
       setDisplayModal(false);
       onDelete();
-    } else {
-      alert("Ajja Bajja, inte din annons!");
     }
   };
 
@@ -36,6 +57,12 @@ export default function DeleteAuctionBtn({ auction, onDelete }) {
           </Modal.Title>
         </Modal.Header>
         <Modal.Body>
+          {/* Display error msg if input-empty */}
+          {error && (
+            <Alert variant="danger" className="mb-2">
+              {error}
+            </Alert>
+          )}
           <p>
             Ange din <strong>privata</strong> nyckel för att radera annonsen.
           </p>
@@ -45,6 +72,7 @@ export default function DeleteAuctionBtn({ auction, onDelete }) {
             value={deleteKey}
             onChange={(e) => setDeleteKey(e.target.value)}
             placeholder="Ange CreatedBy-värdet"
+            disabled={disableInput}
           />
         </Modal.Body>
         <Modal.Footer>
@@ -55,7 +83,12 @@ export default function DeleteAuctionBtn({ auction, onDelete }) {
           >
             Avbryt
           </Button>
-          <Button size="sm" variant="danger" onClick={handleDelete}>
+          <Button
+            size="sm"
+            variant="danger"
+            onClick={handleDelete}
+            disabled={disableInput}
+          >
             Radera
           </Button>
         </Modal.Footer>
